fix(gateway): preserve provider usage error messages
Keep user-friendly ProviderUsageError.message stable and store raw error text in detail.

Deduplicate safeDetail helper across provider usage and usage routes.
